Add files via upload
Here you need to create some documents of different classes and should place in document. Rough code that calculates the probability of the diseases based on the symptoms given in the input.Based on the highest probability it predicts the disease. It basically works on naive bayes and also based on the document classification.